pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "polybrain"
version = "0.0.0"
requires-python = ">=3.12"
authors = [
  {name = "Kyle Tennison", email = "kyletennison@gmail.com"},
]
maintainers = [
    {name = "Kyle Tennison", email = "kyletennison@gmail.com"},
]
description = "An AI powered 3D modeling assistant"
readme = "README.md"
license = {file = "LICENSE"}
classifiers = [
  "Programming Language :: Python"
]
dependencies = [
    "loguru",
    "requests",
    "pydantic",
    "numpy",
    "prettytable",
    "langchain",
    "langchain-openai",
    "python-dotenv",
    "langchainhub",
    "playsound",
    "PyObjC",
    "Levenshtein",
    "fastapi",
    "uvicorn",
    "uvicorn[standard]",
]
[project.optional-dependencies]
dev = [
    "black",
    "pytest"
]
